Add product and reviews controller

+++ controllers/productsController.js
@@ -4,6 +4,7 @@ import Brand from '../models/brand.js';
 const getAllProducts = async (_req, res, next) => {
   try {
     const products = await Product.find();
+
     return res.status(200).json(products);
   } catch (err) {
     next(err);
@@ -24,12 +25,6 @@ async function searchProducts(req, res, next) {
       ]
     });
 
-    // where we dynamically add the search query and the search query value
-    // const { searchBy, value } = req.query;
-    // const products = await Product.find({
-    //   [searchBy]: { $regex: value, $options: 'i' }
-    // });
-
     return res.status(200).json(products);
   } catch (error) {
     next(error);
@@ -40,8 +35,8 @@ const createNewProduct = async (req, res, next) => {
   console.log('CURRENT USER', req.currentUser);
   try {
     const product = await Product.create({
-      ...req.body
-      // addedBy: req.currentUser._id
+      ...req.body,
+      addedBy: req.currentUser._id
     });
 
     await Brand.findOneAndUpdate(
@@ -57,8 +52,9 @@ const createNewProduct = async (req, res, next) => {
 
 const getSingleProduct = async (req, res, next) => {
   try {
-    const product = await Product.findById(req.params.id).populate('brand');
-    // .populate('reviews.reviewer');
+    const product = await Product.findById(req.params.id)
+      .populate('brand')
+      .populate('reviews.reviewer');
     return product
       ? res.status(200).json(product)
       : res
@@ -71,22 +67,17 @@ const getSingleProduct = async (req, res, next) => {
 
 const updateSingleProduct = async (req, res, next) => {
   try {
-    // if (
-    //   req.currentUser._id.equals(product.addedBy) ||
-    //   req.currentUser.isAdmin
-    // ) {
-    //   const product = await Product.findById(req.params.id);
-    //   product.set(req.body);
-    //   const updatedProduct = await product.save();
-    //   return res.status(200).json(updatedProduct);
-    // }
-
-    // return res.status(301).json({ message: 'Unauthorised' });
-
     const product = await Product.findById(req.params.id);
-    product.set(req.body);
-    const updatedProduct = await product.save();
-    return res.status(200).json(updatedProduct);
+    if (
+      req.currentUser._id.equals(product.addedBy) ||
+      req.currentUser.isAdmin
+    ) {
+      product.set(req.body);
+      const updatedProduct = await product.save();
+
+      return res.status(200).json(updatedProduct);
+    }
+    return res.status(301).json({ message: 'Unauthorised' });
   } catch (err) {
     next(err);
   }
@@ -96,20 +87,17 @@ const deleteSingleProduct = async (req, res, next) => {
   try {
     const product = await Product.findById(req.params.id);
 
-    // console.log(product.addedBy, req.currentUser._id);
+    console.log(product.addedBy, req.currentUser._id);
 
-    // if (
-    //   req.currentUser._id.equals(product.addedBy) ||
-    //   req.currentUser.isAdmin
-    // ) {
-    //   await Product.findByIdAndDelete(req.params.id);
-    //   return res.status(200).json({ message: 'Successfully deleted product' });
-    // }
+    if (
+      req.currentUser._id.equals(product.addedBy) ||
+      req.currentUser.isAdmin
+    ) {
+      await Product.findByIdAndDelete(req.params.id);
+      return res.status(200).json({ message: 'Successfully deleted product' });
+    }
 
-    // return res.status(301).json({ message: 'Unauthorised' });
-
-    await Product.findByIdAndDelete(req.params.id);
-    return res.status(200).json({ message: 'Successfully deleted product' });
+    return res.status(301).json({ message: 'Unauthorised' });
   } catch (err) {
     next(err);
   }
